Dev to main

+++ src/waterpump.h
@@ -7,6 +7,6 @@
 void waterpump_init(void); //initaliserer vandpumpen
 void waterpump_on(void);      // Tænd vandpumpen
 void waterpump_off(void);     // Sluk vandpumpen
-void waterpump_run(uint32_t seconds); // Kør pumpen i et bestemt antal sekunder
+void waterpump_run(uint8_t seconds); // Kør pumpen i et bestemt antal sekunder
 
 #endif
